Recording correct receiver of redirected emails
