--- config/initializers/email_redirections.rb
@@ -1,4 +1,0 @@
-# frozen_string_literal: true
-
-require "email_redirector"
-ActionMailer::Base.register_interceptor(EmailRedirector)
