+++ config/initializers/mail.rb
@@ -1,10 +1,9 @@
 # frozen_strin_literal: true
 
 require 'mail/delivery_recorder'
+require "mail/redirector"
 
+
+ActionMailer::Base.register_interceptor(Mail::Redirector)
+Mail::Message.include(Mail::Redirector::MessageExtension)
 ActionMailer::Base.register_observer(Mail::DeliveryRecorder)
-
-
-class << ActionMailer::Base
-  attr_accessor :record_emails
-end
